m1/requirements.txt
pandas
geopandas
shapely
uvicorn
fastapi
python-dotenv
supabase
